backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
hume
python-dotenv==1.0.0
httpx==0.27.0
pydub==0.25.1
google-cloud-storage==2.18.2
google-auth==2.35.0
pydantic==2.5.0
